+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ### 1.8.2 (2019-01-25)
 
 - Clean the cache to avoid issues due to cache files during functional tests
+- Correct Unzip assets command help 
 
 ### 1.8.1 (2019-01-23)